+++ app.js
@@ -25,8 +25,11 @@ SOFTWARE.
 var http = require('http');
 var express = require('express');
 
+var morgan = require('morgan');
+var fileStreamRotator = require('file-stream-rotator');
+var fs = require('fs');
+
 var cors = require('cors');
-var logger = require('morgan');
 var bodyParser = require('body-parser');
 var methodOverride = require('method-override');
 var session = require('express-session');
@@ -48,6 +51,7 @@ program
     'MongoDB database name (default: camomile)')
   .option('--root-password <dbname>', 'Change/set root password')
   .option('--media <dir>', 'Path to media root directory')
+  .option('--log <dir>', 'Path to log directory (default: ' + __dirname + '/log)')
   .parse(process.argv);
 
 var port =
@@ -84,10 +88,55 @@ var media =
   process.env.MEDIA ||
   '/media';
 
-var cookieSecret = process.env.COOKIE_SECRET || Authentication.helper.cookieSecret();
+// CLI || env || ./log
+var logDirectory =
+  program.log ||
+  process.env.LOG ||
+  __dirname + '/log';
 
-mongoose.connect('mongodb://' + mongodb_host + ':' + mongodb_port + '/' +
-  mongodb_name);
+// env || random
+var cookieSecret =
+  process.env.COOKIE_SECRET ||
+  Authentication.helper.cookieSecret();
+
+var app = express();
+
+app.set('port', port);
+app.set('media', media);
+
+// === LOGGING ================================================================
+
+// create log directory
+fs.existsSync(logDirectory) || fs.mkdirSync(logDirectory);
+
+// create a (daily) rotating write stream
+var accessLogStream = fileStreamRotator.getStream({
+  filename: logDirectory + '/%DATE%.log',
+  frequency: 'daily',
+  verbose: false,
+  date_format: 'YYYYMMDD'
+});
+
+morgan.token('user', function (req, res) {
+  if (req.session && req.session.user) {
+    return req.session.user.username;
+  } else {
+    return 'anonymous';
+  };
+});
+
+var logFormat = '[:date[clf]] :user (:remote-addr) :method :url :status :response-time ms';
+
+var logger = morgan(logFormat, {
+  skip: function (req, res) {
+    return req.method === 'GET' || req.method === 'OPTIONS';
+  },
+  stream: accessLogStream
+});
+
+app.use(logger);
+
+// === CROSS-ORIGIN RESOURCE SHARING ==========================================
 
 var cors_options = {
   origin: true,
@@ -99,19 +148,23 @@ var cors_options = {
   credentials: true,
 };
 
+app.use(cors(cors_options));
+
+// ============================================================================
+
+app.use(methodOverride());
+
+// === SESSION ================================================================
+
+mongoose.connect('mongodb://' + mongodb_host + ':' + mongodb_port + '/' +
+  mongodb_name);
+
 var sessionStore = new mongoStore({
   mongooseConnection: mongoose.connection,
   db: mongoose.connections[0].db,
   clear_interval: 60
 });
 
-var app = express();
-
-app.set('port', port);
-app.set('media', media);
-app.use(logger('dev'));
-app.use(cors(cors_options));
-app.use(methodOverride());
 app.use(session({
   cookie: {
     secure: false,
@@ -120,13 +173,16 @@ app.use(session({
   name: 'camomile.sid',
   proxy: true,
   resave: false,
-  rolling: false,
+  rolling: true,
   saveUninitialized: false,
   secret: cookieSecret,
   store: sessionStore,
 }));
 
 app.use(cookieParser(cookieSecret));
+
+// ============================================================================
+
 app.use(bodyParser.json({
   limit: '50mb'
 }));
@@ -134,6 +190,8 @@ app.use(bodyParser.urlencoded({
   limit: '50mb',
   extended: true
 }));
+
+// ============================================================================
 
 // handle CORS pre-flight requests
 // (must be added before any other route)